Refactor: 패키지 구조 변경
- 회원가입 및 로그인 API를 기존 'user' 패키지에서 'auth' 패키지로 이동
