+++ backend/src/main/java/com/sungkyunkwan/tload/domain/auth/controller/AuthController.java
@@ -1,4 +1,4 @@
-package com.sungkyunkwan.tload.domain.user.controller;
+package com.sungkyunkwan.tload.domain.auth.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,11 +7,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.sungkyunkwan.tload.domain.user.dto.SigninRequestDto;
-import com.sungkyunkwan.tload.domain.user.dto.SigninResponseDto;
-import com.sungkyunkwan.tload.domain.user.dto.SignupRequestDto;
-import com.sungkyunkwan.tload.domain.user.dto.SignupResponseDto;
-import com.sungkyunkwan.tload.domain.user.service.UserService;
+import com.sungkyunkwan.tload.domain.auth.dto.SigninRequestDto;
+import com.sungkyunkwan.tload.domain.auth.dto.SigninResponseDto;
+import com.sungkyunkwan.tload.domain.auth.dto.SignupRequestDto;
+import com.sungkyunkwan.tload.domain.auth.dto.SignupResponseDto;
+import com.sungkyunkwan.tload.domain.auth.service.AuthService;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -19,17 +19,17 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/auth")
-public class UserController {
+public class AuthController {
 
-	private final UserService userService;
+	private final AuthService authService;
 
 	@PostMapping("/signup")
 	public ResponseEntity<SignupResponseDto> signup(@Valid @RequestBody SignupRequestDto signupRequestDto) {
-		return ResponseEntity.status(HttpStatus.CREATED).body(userService.signup(signupRequestDto));
+		return ResponseEntity.status(HttpStatus.CREATED).body(authService.signup(signupRequestDto));
 	}
 
 	@PostMapping("/login")
 	public ResponseEntity<SigninResponseDto> login(@Valid @RequestBody SigninRequestDto signinRequestDto) {
-		return ResponseEntity.status(HttpStatus.CREATED).body(userService.login(signinRequestDto));
+		return ResponseEntity.status(HttpStatus.CREATED).body(authService.login(signinRequestDto));
 	}
 }
