+++ backend/src/main/java/com/sungkyunkwan/tload/domain/auth/dto/SigninResponseDto.java
@@ -1,4 +1,4 @@
-package com.sungkyunkwan.tload.domain.user.dto;
+package com.sungkyunkwan.tload.domain.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
