+++ backend/src/main/java/com/sungkyunkwan/tload/domain/auth/dto/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package com.sungkyunkwan.tload.domain.user.dto;
+package com.sungkyunkwan.tload.domain.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
